refactor: 페이크 객체 패키지 이동
