+++ src/test/java/roomescape/service/InMemoryReservationTimeDao.java
@@ -1,8 +1,9 @@
-package roomescape.dao.resetvationTime;
+package roomescape.service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.concurrent.atomic.AtomicLong;
+import roomescape.dao.resetvationTime.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 
 public class InMemoryReservationTimeDao implements ReservationTimeDao {
