+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -7,9 +7,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.dao.reservation.InMemoryReservationDao;
 import roomescape.dao.reservation.ReservationDao;
-import roomescape.dao.resetvationTime.InMemoryReservationTimeDao;
 import roomescape.dao.resetvationTime.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.request.ReservationCreateRequest;
